requirements.txt
streamlit
openai
pymupdf
pytesseract
sentence-transformers
chromadb
python-dotenv
pillow
langchain
plotly
umap-learn
rouge
nltk
rouge-score
pytorch
